requirements.txt
 pandas
 numpy
 os
 pickle
 tqdm.notebook
 PIL
 tensorflow
 cv2
 matplotlib
 unicodedata
 streamlit